cubic merge

See merge request VoiceInYourHead/manhattan-city!39

+++ code/modules/mob/mob.dm
@@ -235,17 +235,20 @@
 		return 0
 
 	var/tile = get_turf(A)
-	if (!tile)
+	if(!tile)
 		return 0
 
-	var/obj/P = new /obj/effect/decal/point(tile)
+	var/obj/effect/decal/point/P = new(tile)
 	P.invisibility = invisibility
 	P.plane = plane
+	P.pixel_x = A.pixel_x
+	P.pixel_y = A.pixel_y
+	
 	spawn (20)
-		if(P)
-			qdel(P)	// qdel
+		P && qdel(P) // qdel
 
 	face_atom(A)
+
 	return 1
